inflearn-45 OSIV 와 성능 최적화 - 서비스 로직 분리
