+++ src/main/java/jpabook/jpashop/service/query/OrderItemDto.java
@@ -1,0 +1,19 @@
+package jpabook.jpashop.service.query;
+
+import jpabook.jpashop.domain.OrderItem;
+import lombok.Getter;
+
+@Getter
+public class OrderItemDto {
+
+
+    private String itemName;    // 상품 명
+    private int orderPrice;     // 주문 가격
+    private int count;          // 주문 수량
+
+    public OrderItemDto(OrderItem orderItem) {
+        itemName = orderItem.getItem().getName();
+        orderPrice = orderItem.getOrderPrice();
+        count = orderItem.getCount();
+    }
+}
