+++ src/main/resources/application.yml
@@ -12,7 +12,7 @@ spring:
 #       show_sql: true system.out 을 통해 표시
         format_sql: true
         default_batch_fetch_size: 100
-
+    open-in-view: false
 logging:
   level:
     org.hibernate.SQL: debug
